[Clusters] Change TPU INFO log to debug and only query TPU endpoint if TPUs are detected (#40079)
Currently, a requests.get() call is made to a GCE endpoint every time ray.init() is called. It's only supposed to be called when TPUs are detected, but it was in fact being called every time.

This is the main issue that this PR fixes. We now only run the accelerator-specific code if the number of accelerators detected is > 0. (Previously we were only checking that it was not None.)

Another issue this PR fixes is that a "TPUs not detected" message was being printed to the driver in some cases (e.g. when using Ray Serve, see the linked issue.). We fix this by setting it as a debug-level log instead of a info-level log. I don't know a user-friendly way to enable the debug log level, but we can improve this in the future.

This PR also adds a 30s timeout to the requests.get call, which is a best practice (otherwise it can silently hang forever.)

Related issue number
Closes #40078

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>