[Serve] add some logging to `test_standalone_3.py` (#40120)
For some reason, I'm unable to reproduce the issue after re-running this 7 times. My hypothesis is there are some things in windows causing the operation not complete in the sequence we expect and that adding the logging caused some time delayed inbetween. So when the assertion ran, it actually have the operation completed and checked the correct state. If this doesn't solve the root cause, it can also help to debug future issue if it comes up again.